requirements.txt
fastapi
paho-mqtt
motor
aioredis
matplotlib
uvicorn